+++ lib/src/view/uxutils/src/view/nav_bottom_bar.dart
@@ -1,3 +1,4 @@
+//
 /// This BottomNavigationBar wrapper is no advantage.
 /// BottomNavigationBar is a StatefulWidget!
 /// Merely assign to a variable and then you're free to update its items.